WIP: fixing bugs in new init
This is taking a while and may not be worth the effort. We will be using
SystemD to manage services. This will provide most of the same
advantages that the new init system provides while also simplifying and
reducing our support burden.

So, I'm creating this branch with the work done so far and will try to
instead simply patch the older init scripts to work as SystemD units.