Primary MailCleaner target
Starts and stops all services correctly. However, the mariadb services
take ~10 seconds to stop. If I try to start before they fully stop, it
will fail because most of the other service are dependent on the DBs
which will have stopped or be stopping.

It is possible to add a delay before the mailcleaner target exits, but
it would be much better if it actually waits for all services to stop
before returning. Still reading the docs for this...