ClamAV has issues with AppArmor
This creates a local apparmor directory/configuration file which allows
it to have access to our directories. It also creates a symlink to this
directory if one does not exist in each of the major service dumpers.

It is done this way so that we can restrict other services further in
the future also.

Not fully working, so I'm commiting as a new branch.