libipc-run3-perl != IPC::Run
It lacks pipe and redirect support, which is exactly what we got
IPC::Run for. IPC::Run is not available as a deb, so well have to get it
from CPAN.

For now, we will install with MailScanner installation.

TODO: Separate script, or integration in install/install.sh to install
CPAN packages. Eventually MailScanner will hopefully be containerized,
so we cannot depend on it's library list in the base OS.